modify all key pages, modify middlesection part

+++ app/Http/Controllers/UserController.php
@@ -7,60 +7,49 @@ use App\HDtype;
 use App\HelpDesk;
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Crypt;
 use Illuminate\Support\Facades\Hash;
 use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
-        $user=User::all();
-        $type=HDtype::all();
-        $priority = HDpriority::ALL();
-        $help_desk = HelpDesk::where('hhd_ticket_status','1')->get();
+
+        $current_user = auth()->user()->id;
+        $help_desk = HelpDesk::select('hhd_request_user_id', 'id', 'hhd_type', 'hhd_priority')->where('hhd_ticket_status', '1')->where('hhd_receiver_user_id', $current_user)->get();
+        $type = HDtype::select('th_name', 'id')->get();
+        $priority = HDpriority::select('id', 'hdp_name')->get();
+        $user = User::select('id', 'name')->get();
         $users = User::all();
-        $role_user=User::where('id',auth()->user()->id)->get()->first();
-        return view('users.index',compact('role_user','help_desk','type','priority','user'))->with('users', $users);
+        $role_user = User::where('id', auth()->user()->id)->get()->first();
+        return view('users.index', compact('role_user', 'help_desk', 'type', 'priority', 'user'))->with('users', $users);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
-        $user=User::all();
-        $type=HDtype::all();
-        $priority = HDpriority::ALL();
-        $help_desk = HelpDesk::where('hhd_ticket_status','1')->get();
+        $current_user = auth()->user()->id;
+        $help_desk = HelpDesk::select('hhd_request_user_id', 'id', 'hhd_type', 'hhd_priority')->where('hhd_ticket_status', '1')->where('hhd_receiver_user_id', $current_user)->get();
+        $type = HDtype::select('th_name', 'id')->get();
+        $priority = HDpriority::select('id', 'hdp_name')->get();
+        $user = User::select('id', 'name')->get();
         //Get all roles and pass it to the view
         $roles = Role::get();
-        return view('users.create', ['roles'=>$roles],compact('help_desk','priority','type','user'));
+        return view('users.create', ['roles' => $roles], compact('help_desk', 'priority', 'type', 'user'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
 
         //Validate name, email and password fields
         $this->validate($request, [
-            'name'=>'required|max:120',
+            'name' => 'required|max:120',
 //            'email'=>'required|email|unique:users',
-            'password'=>'required|min:6|confirmed'
+            'password' => 'required|min:6|confirmed'
         ]);
 
-        $user = User::create($request->only( 'name', 'password','username','device_id')); //Retrieving only the email and password data
+        $user = User::create($request->only('name', 'password', 'username', 'device_id')); //Retrieving only the email and password data
         $roles = $request['roles']; //Retrieving the roles field
         //Checking if a role was selected
         if (isset($roles)) {
@@ -75,79 +64,60 @@ class UserController extends Controller
                 'User successfully added.');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
-        $user=User::all();
-        $type=HDtype::all();
-        $priority = HDpriority::ALL();
-        $help_desk = HelpDesk::where('hhd_ticket_status','1')->get();
-        return redirect('users','help_desk','priority','type');
+        $current_user = auth()->user()->id;
+        $help_desk = HelpDesk::select('hhd_request_user_id', 'id', 'hhd_type', 'hhd_priority')->where('hhd_ticket_status', '1')->where('hhd_receiver_user_id', $current_user)->get();
+        $type = HDtype::select('th_name', 'id')->get();
+        $priority = HDpriority::select('id', 'hdp_name')->get();
+        $user = User::select('id', 'name')->get();
+        return redirect('users', 'help_desk', 'priority', 'type');
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
-        $user=User::all();
-        $type=HDtype::all();
-        $priority = HDpriority::ALL();
-        $help_desk = HelpDesk::where('hhd_ticket_status','1')->get();
+
+        $current_user = auth()->user()->id;
+        $help_desk = HelpDesk::select('hhd_request_user_id', 'id', 'hhd_type', 'hhd_priority')->where('hhd_ticket_status', '1')->where('hhd_receiver_user_id', $current_user)->get();
+        $type = HDtype::select('th_name', 'id')->get();
+        $priority = HDpriority::select('id', 'hdp_name')->get();
+        $user = User::select('id', 'name')->get();
         $user = User::findOrFail($id); //Get user with specified id
         $roles = Role::get(); //Get all roles
 
-        return view('users.edit', compact('user', 'roles','help_desk','priority','type')); //pass user and roles data to view
+        return view('users.edit', compact('decrypt', 'user', 'roles', 'help_desk', 'priority', 'type')); //pass user and roles data to view
 
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, $id)
     {
 
 
         $user = User::findOrFail($id); //Get role specified by id
-
         //Validate name, email and password fields
         $this->validate($request, [
 //            'name'=>'required|max:120',
 //            'password'=>'required|min:6|confirmed'
         ]);
-        $input = $request->only(['name', 'password','username','device_id']); //Retreive the name, email and password fields
-        $roles = $request['roles']; //Retreive all roles
-        $user->fill($input)->save();
+        $result = Hash::check($request->password, $user->password);
+        if ($result == true) {
+            $input = $request->only(['name', 'newPassword']); //Retreive the name, email and password fields
+            $roles = $request['roles']; //Retreive all roles
+            $user->name = $request->name;
+            $user->password = $request->newPassword;
+            $user->save();
 
-        if (isset($roles)) {
-            $user->roles()->sync($roles);  //If one or more role is selected associate user to roles
+            if (isset($roles)) {
+                $user->roles()->sync($roles);  //If one or more role is selected associate user to roles
+            } else {
+                $user->roles()->detach(); //If no role is selected remove exisiting role associated to a user
+            }
+            return redirect()->route('users.index')
+                ->with('flash_message',
+                    'User successfully edited.');
         }
-        else {
-            $user->roles()->detach(); //If no role is selected remove exisiting role associated to a user
-        }
-        return redirect()->route('users.index')
-            ->with('flash_message',
-                'User successfully edited.');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
         $user = User::findOrFail($id);
@@ -175,7 +145,7 @@ class UserController extends Controller
 
         $data = '';
         foreach ($order as $orders) {
-            $data .= '["' . $orders->id . '",' . '"' . $orders->hp_project_name . '",' . '"' . $orders->hp_employer_name . '",' . '"' . $orders->hp_connector . '",' . '"' . $orders->hp_type_project. '"],';
+            $data .= '["' . $orders->id . '",' . '"' . $orders->hp_project_name . '",' . '"' . $orders->hp_employer_name . '",' . '"' . $orders->hp_connector . '",' . '"' . $orders->hp_type_project . '"],';
         }
         $data = substr($data, 0, -1);
         $orders_count = Order::all()->count();
